Existing xpath is not supported in ruby -v 2.4
Unable to use the Xpath in ruby version 2.4. It returns only one data in the array as tag <xenc:CipherValue></xenc:CipherValue>.
we need <e:CipherValue> in to perform authenticate function. it throws an error as 'text function is call by an nil object.'
line 103 : @security_token1 = cipher_values[1].text